// src/app/components/model/vacinados.ts
export interface Vacinados {
  id: number;
  document_id: string;
  paciente_id: string;
  paciente_idade: string;
  paciente_datanascimento: string;
  paciente_enumsexobiologico: string;
  paciente_racacor_codigo: string;
  paciente_racacor_valor: string;
  paciente_endereco_coibgemunicipio: string;
  paciente_endereco_nmmunicipio: string;
  paciente_endereco_nmpais: string;
  paciente_endereco_uf: string;
  paciente_endereco_cep: string;
  paciente_nacionalidade_enumnacionalidade: string;
  estabelecimento_valor: string;
  estabelecimento_razaosocial: string;
  estabelecimento_nofantasia: string;
  estabelecimento_municipio_codigo: string;
  estabelecimento_municipio_nome: string;
  estabelecimento_uf: string;
  vacina_grupoatendimento_codigo: string;
  vacina_grupoatendimento_nome: string;
  vacina_categoria_codigo: string;
  vacina_categoria_nome: string;
  vacina_lote: string;
  vacina_fabricante_nome: string;
  vacina_dataaplicacao: string;
  vacina_descricao_dose: string;
  vacina_codigo: string;
  vacina_nome: string;
}